# JIT: Optimize SequenceEqual to use ccmp on ARM64