In the original PR we could not get this this working due to some conservative interference. This now does the right thing with #92710 merged.

Also change LowerCallMemcmp/LowerCallMemmove to return next node to lower just to align it a bit more with other functions.

```csharp
bool Test1(ReadOnlySpan<byte> data) => "hello world"u8.SequenceEqual(data);
```

Before:
```asm
G_M52062_IG01:  ;; offset=0x0000
            stp     fp, lr, [sp, #-0x10]!
            mov     fp, sp
						;; size=8 bbWeight=1 PerfScore 1.50
G_M52062_IG02:  ;; offset=0x0008
            movz    x2, #0x20D4      // data for <PrivateImplementationDetails>:430646847E70344C09F58739E99D5BC96EAC8D5FE7295CF196B986279876BF9B
            movk    x2, #0xE179 LSL #16
            movk    x2, #554 LSL #32
            cmp     w1, #11
            bne     G_M52062_IG05
						;; size=20 bbWeight=1 PerfScore 3.00
G_M52062_IG03:  ;; offset=0x001C
            ldr     x1, [x2]
            ldr     x2, [x2, #0x03]
            ldr     x3, [x0]
            ldr     x0, [x0, #0x03]
            eor     x1, x1, x3
            eor     x0, x2, x0
            orr     x0, x1, x0
            cmp     x0, #0
            cset    x0, eq
						;; size=36 bbWeight=0.48 PerfScore 6.96
G_M52062_IG04:  ;; offset=0x0040
            ldp     fp, lr, [sp], #0x10
            ret     lr
						;; size=8 bbWeight=1 PerfScore 2.00
G_M52062_IG05:  ;; offset=0x0048
            mov     w0, wzr
            b       G_M52062_IG04
						;; size=8 bbWeight=0 PerfScore 0.00
```

After:
```asm
G_M52062_IG01:  ;; offset=0x0000
            stp     fp, lr, [sp, #-0x10]!
            mov     fp, sp
						;; size=8 bbWeight=1 PerfScore 1.50
G_M52062_IG02:  ;; offset=0x0008
            movz    x2, #0x20D4      // data for <PrivateImplementationDetails>:430646847E70344C09F58739E99D5BC96EAC8D5FE7295CF196B986279876BF9B
            movk    x2, #0x63E LSL #16
            movk    x2, #445 LSL #32
            cmp     w1, #11
            bne     G_M52062_IG05
						;; size=20 bbWeight=1 PerfScore 3.00
G_M52062_IG03:  ;; offset=0x001C
            ldr     x1, [x2]
            ldr     x2, [x2, #0x03]
            ldr     x3, [x0]
            ldr     x0, [x0, #0x03]
            cmp     x1, x3
            ccmp    x2, x0, 0, eq
            cset    x0, eq
						;; size=28 bbWeight=0.48 PerfScore 6.48
G_M52062_IG04:  ;; offset=0x0038
            ldp     fp, lr, [sp], #0x10
            ret     lr
						;; size=8 bbWeight=1 PerfScore 2.00
G_M52062_IG05:  ;; offset=0x0040
            mov     w0, wzr
            b       G_M52062_IG04
						;; size=8 bbWeight=0 PerfScore 0.00
```

Benchmarks from #83945 on my M1:
|       Method |        Job |               Toolchain |      Mean |     Error |    StdDev | Ratio |
|------------- |----------- |------------------------ |----------:|----------:|----------:|------:|
|      TE_Json | Job-RFUZAM |      /Core_Root/corerun | 1.9945 ns | 0.0100 ns | 0.0083 ns |  0.98 |
|      TE_Json | Job-KSXFQI | /Core_Root_base/corerun | 2.0339 ns | 0.0011 ns | 0.0009 ns |  1.00 |
|              |            |                         |           |           |           |       |
| TE_Plaintext | Job-RFUZAM |      /Core_Root/corerun | 0.7460 ns | 0.0086 ns | 0.0080 ns |  0.99 |
| TE_Plaintext | Job-KSXFQI | /Core_Root_base/corerun | 0.7515 ns | 0.0008 ns | 0.0008 ns |  1.00 |
|              |            |                         |           |           |           |       |
|    Equals_15 | Job-RFUZAM |      /Core_Root/corerun | 0.5063 ns | 0.0008 ns | 0.0008 ns |  0.87 |
|    Equals_15 | Job-KSXFQI | /Core_Root_base/corerun | 0.5799 ns | 0.0006 ns | 0.0006 ns |  1.00 |